Fix parallel parking bad clear

+++ planning/behavior_path_planner/include/behavior_path_planner/scene_module/pull_over/pull_over_module.hpp
@@ -90,7 +90,7 @@ struct PullOverArea
 {
   Pose start_pose;
   Pose end_pose;
-  float distance_from_goal;
+  double distance_from_goal;
 
   Pose center_pose = calcAveragePose(start_pose, end_pose);
